test(ffmpeg): split HW crop tests into per-hardware build tags
Replace the single hwtest crop tests with dedicated build tags and files:
- transcode_vaapi_test.go (vaapitest): TestTranscode_CropWithVAAPI_SWDecode
- transcode_nvenc_test.go (nvenctest): TestTranscode_CropWithNVENC_CuvidFallback

transcode_hw_test.go retains only TestDetectHardwareEncoders_HardwarePresent
(which validates generic HW detection, not a specific accelerator).

Makefile gains HAS_VAAPI_ENCODER and HAS_NVENC_ENCODER detection via ffmpeg
-encoders, following the existing HAS_QSV_ENCODER pattern. The _test_tags
variable is updated to include vaapitest and nvenctest when the corresponding
hardware is present.
