+++ pkg/ffmpeg/transcode_nvenc_test.go
@@ -1,0 +1,48 @@
+//go:build nvenctest
+
+package ffmpeg_test
+
+import (
+	"path/filepath"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/solidDoWant/media-processor/pkg/ffmpeg"
+	"github.com/solidDoWant/media-processor/pkg/ffprobe"
+)
+
+// TestTranscode_CropWithNVENC_CuvidFallback verifies that the CUDA cuvid
+// fallback crop path (hwdownload,crop=W:H:X:Y,hwupload) produces output with
+// the correct dimensions when the cuvid dict option is unavailable or unsupported
+// for the input codec.
+// Requires NVENC hardware (nvenctest build tag).
+func TestTranscode_CropWithNVENC_CuvidFallback(t *testing.T) {
+	output := filepath.Join(t.TempDir(), "out.mkv")
+	crop := &ffmpeg.CropParams{W: 320, H: 176, X: 0, Y: 22}
+
+	err := ffmpeg.NewTranscode(testBlackBarsVideoPath, output).
+		ToVideoCodec(ffmpeg.CodecH265).
+		ToContainer(ffmpeg.ContainerMKV).
+		HardwareAccel(ffmpeg.HWAccelNVENC).
+		WithCrop(crop).
+		Build().
+		Run(t.Context())
+	require.NoError(t, err)
+
+	info, err := ffprobe.Probe(t.Context(), output)
+	require.NoError(t, err)
+
+	var videoStream *ffprobe.StreamInfo
+	for i := range info.Streams {
+		if info.Streams[i].CodecType == ffprobe.CodecTypeVideo {
+			videoStream = &info.Streams[i]
+			break
+		}
+	}
+
+	require.NotNil(t, videoStream, "output must contain a video stream")
+	assert.Equal(t, crop.W, videoStream.WidthPixels, "output width must match crop width")
+	assert.Equal(t, crop.H, videoStream.HeightPixels, "output height must match crop height")
+}
